+++ src/ethereum_types/v0_15.rs
@@ -17,7 +17,7 @@ macro_rules! impl_varint_for {
         pub const fn [<encode_ $ty:snake _to>](
           val: &$ty,
           buf: &mut [u8],
-        ) -> Result<usize, $crate::EncodeError> {
+        ) -> Result<usize, $crate::ConstEncodeError> {
           $crate::bnum::encode_uint_d64_to($target::from_digits(val.0), buf)
         }
 
@@ -25,7 +25,7 @@ macro_rules! impl_varint_for {
         ///
         /// Returns the bytes read and the value.
         #[inline]
-        pub const fn [<decode_ $ty:snake>](buf: &[u8]) -> Result<(usize, $ty), $crate::DecodeError> {
+        pub const fn [<decode_ $ty:snake>](buf: &[u8]) -> Result<(usize, $ty), $crate::ConstDecodeError> {
           match $crate::bnum::decode_uint_d64(buf) {
             Ok((read, val)) => Ok((read, $ty(*val.digits()))),
             Err(e) => Err(e),
@@ -42,13 +42,13 @@ macro_rules! impl_varint_for {
           }
 
           fn encode(&self, buf: &mut [u8]) -> Result<usize, $crate::EncodeError> {
-            [<encode_ $ty:snake _to>](self, buf)
+            [<encode_ $ty:snake _to>](self, buf).map_err(Into::into)
           }
 
           fn decode(buf: &[u8]) -> Result<(usize, Self), $crate::DecodeError>
             where
               Self: Sized {
-            $target::decode(buf).map(|(len, value)| (len, $ty(value.into())))
+            $target::decode(buf).map(|(len, value)| (len, $ty(value.into()))).map_err(Into::into)
           }
         }
       )*